Add auto formatting for neovim
This revealed some problems with how I was using my bashrc. I had to use
an interactive shell in neovim to access my custom PATH, but that should
really be handled in my bash profile. I've updated that, so that now the
bash profile is also tracked, and will handle all PATH changes